// src/api/models/Group.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Message } from './Message';
export type Group = {
    id?: string;
    name?: string;
    iconUrl?: string | null;
    memberCount?: number;
    unreadCount?: number;
    lastMessage?: Message;
    createdAt?: string;
    updatedAt?: string;
};